web/client: List table component refactoring
