+++ web/client/src/lib/components/List/Content/Table/ContentTableHead/index.ts
@@ -1,0 +1,3 @@
+import ContentTableHead from './ContentTableHead';
+
+export default ContentTableHead;
