+++ web/client/src/lib/components/List/Content/Table/ContentTableRow/ContentTableRow.styles.tsx
