+++ web/client/src/lib/components/List/Content/Table/ContentTableRow/index.ts
@@ -1,0 +1,3 @@
+import ContentTableRow from './ContentTableRow';
+
+export default ContentTableRow;
